emp-portal/EMApp/app/requirements.txt
alabaster==0.7.12
asgiref==3.2.7
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==2.2.14
django-cors-headers==3.3.0
django-discover-runner==1.0
django-elasticsearch-dsl==7.1.1
django-markupfield==2.0.0
django-model-utils==4.0.0
django-nine==0.2.3
django-phone-field==1.8.0
django-redis==4.12.1
django-role-permissions==3.0.0
django-taggit==1.3.0
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
dnspython==1.16.0
docutils==0.16
elasticsearch==7.6.0
elasticsearch-dsl==7.2.0
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
Jinja2==3.0.0a1
packaging==20.4
Pillow==7.1.2
psycopg2==2.8.5
pyfiglet==0.8.post1
Pygments==2.6.1
pyparsing==3.0.0a1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
django-yamlconf==1.2.1
uWSGI==2.0.19.1
jsonschema==3.2.0
validators==0.15.0
pycountry==20.7.3
email-validator==1.1.1